remove extra blank lines in chapter2
The git help command is used to show you all the documentation shipped with Git about any command. While we’re giving a rough overview of most of the more popular ones in this appendix, for a full listing of all of the possible options and flags for every command, you can always run git help <command>.

We introduced the git help command in Getting Help and showed you how to use it to find more information about the git shell in Setting Up the Server.

prev | next